+++ src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Header from "./Header/Header";
 import AboutUs from "./AboutUs/AboutUs";
-import Story from "./OurStory/OurStory";
+import Story from "./Story/Story";
 import Archive_n_Project from "./Archive_n_Project/Archive_n_Project";
 import Team from "./Team/Team";
 import Gallery from "./Gallery/Gallery";
